feat: dynamic per-score Open Graph image
- src/lib/og.ts: buildOgQuery(state, lng) → /api/og query string
- src/app/api/og/route.tsx: edge ImageResponse, cream palette, cache headers
- src/app/HomeClient.tsx: renamed from page.tsx (keep 'use client')
- src/app/page.tsx: server wrapper + generateMetadata (Next 16 await searchParams)
- NotoSerifSC-sub.otf: 22 KB subset (U+0020-007E,U+00B7,U+5206)
  Font downloaded from: https://github.com/notofonts/noto-cjk/releases/download/Serif2.003/14_NotoSerifSC.zip
  (SubsetOTF/SC/NotoSerifSC-Medium.otf, the main-branch raw URL 404s due to Git LFS)
- tests/og.test.ts: 2 tests, TDD red→green
